refactor basic project structure
