+++ src/__test__/CountriesList.test.js
@@ -1,8 +1,8 @@
 import renderer from 'react-test-renderer';
 import { MemoryRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from '../store';
-import CountriesList from './CountriesList';
+import store from '../redux/store';
+import CountriesList from '../components/countries/CountriesList';
 
 describe('Render CountriesList correctly', () => {
   it('Render CountriesList Screen correctly', () => {
