+++ src/__test__/Details.test.js
@@ -1,15 +1,15 @@
 import renderer from 'react-test-renderer';
 import { MemoryRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from '../features/store';
-import NavBar from './NavBar';
+import store from '../redux/store';
+import Details from '../components/countries/Details';
 
-describe('Render NavBar correctly', () => {
-  it('Render NavBar Screen correctly', () => {
+describe('Render Details correctly', () => {
+  it('Render Details Screen correctly', () => {
     const tree = renderer.create(
       <Provider store={store}>
         <MemoryRouter>
-          <NavBar />
+          <Details />
         </MemoryRouter>
       </Provider>,
     ).toJSON();
